+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ### vNext
 
-* TODO
+* Fixes a bug where schemas with scalars could not be merged when passed to
+  `mergeSchemas` as a string or `GraphQLSchema`.  <br/>
+  [@hayes](https://github.com/hayes) in [#1062](https://github.com/apollographql/graphql-tools/pull/1062)
 
 ### 4.0.4